requirements.txt
streamlit
opencv-python
openai
fpdf
python-dotenv
Pillow
moviepy